# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Api",
        "description": "Lumen Swagger Api",
        "license": {
            "name": "my api"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Api Documentations my api"
        }
    ],
    "paths": {
...
        "/resource": {
            "get": {
                "tags": [
                    "Resource"
                ],
                "summary": "List resource",
                "description": "Returns data of resources",
                "operationId": "List",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "authorization": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Resource"
                ],
                "summary": "Store resource",
                "operationId": "Store",
                "requestBody": {
                    "description": "Input data format",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "owner",
                                    "patrimonial_id",
                                    "type_id",
                                    "remark"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nombre de el recurso",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Descripcion del recurso",
                                        "type": "string"
                                    },
                                    "owner": {
                                        "description": "Dueño",
                                        "type": "integer"
                                    },
                                    "patrimonial_id": {
                                        "description": "Id patrimonio",
                                        "type": "string"
                                    },
                                    "type_id": {
                                        "description": "Tipo de recurso",
                                        "type": "integer"
                                    },
                                    "remark": {
                                        "description": "Remark",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "authorization": []
                    }
                ]
            }
        },
        "/resource/{id}": {
            "get": {
                "tags": [
                    "Resource"
                ],
                "summary": "Show resource",
                "description": "Returns data of resource",
                "operationId": "Show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "resource id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "authorization": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Resource"
                ],
                "summary": "Update resource",
                "operationId": "Update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "resource id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "owner",
                                    "patrimonial_id",
                                    "type_id",
                                    "remark"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nombre de el recurso",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Descripcion del recurso",
                                        "type": "string"
                                    },
                                    "owner": {
                                        "description": "Dueño",
                                        "type": "integer"
                                    },
                                    "patrimonial_id": {
                                        "description": "Id patrimonio",
                                        "type": "string"
                                    },
                                    "type_id": {
                                        "description": "Tipo de recurso",
                                        "type": "integer"
                                    },
                                    "remark": {
                                        "description": "Remark",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "authorization": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Resource"
                ],
                "summary": "Delete existing resource",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteresource",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "resource id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "authorization": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "authorization": {
                "type": "apiKey",
                "name": "authorization",
                "in": "header"
            }
        }
    }
}